Engine Merging //UE4/Dev-Main to Dev-Niagara (//UE4/Dev-Niagara)
#rb none
#test tests pass on PC other than ribbon UV's which is known to need an image blessing

[CL 4320203 by Shaun Kime in Dev-Niagara branch]

+++ Engine/Source/Editor/BlueprintGraph/Classes/K2Node_InputTouchEvent.h
@@ -20,13 +20,13 @@ class UK2Node_InputTouchEvent : public UK2Node_Event
 	TEnumAsByte<EInputEvent> InputKeyEvent;
 
 	UPROPERTY()
-	uint32 bConsumeInput:1;
+	uint8 bConsumeInput:1;
 
 	UPROPERTY()
-	uint32 bExecuteWhenPaused:1;
+	uint8 bExecuteWhenPaused:1;
 
 	UPROPERTY()
-	uint32 bOverrideParentBinding:1;
+	uint8 bOverrideParentBinding:1;
 
 	//~ Begin UK2Node Interface
 	virtual UClass* GetDynamicBindingClass() const override;
